Remove redudant code
