+++ variant/tokenizers/__init__.py
@@ -22,3 +22,4 @@ from .reference_sequence import ReferenceSequence  # noqa: F401
 from .amino_acid_substitution import AminoAcidSubstitution  # noqa: F401
 from .polypeptide_truncation import PolypeptideTruncation  # noqa: F401
 from .silent_mutation import SilentMutation  # noqa: F401
+from .polypeptide_sequence_variant_base import PolypeptideSequenceVariantBase  # noqa: F401, E501
