+++ variant/tokenizers/polypeptide_truncation.py
@@ -1,23 +1,12 @@
 """A module for Polypeptide Truncation Tokenization."""
-import re
 from typing import Optional
-from .caches import AminoAcidCache
-from .tokenizer import Tokenizer
 from variant.schemas.token_response_schema import PolypeptideTruncationToken,\
     TokenMatchType
+from .polypeptide_sequence_variant_base import PolypeptideSequenceVariantBase
 
 
-class PolypeptideTruncation(Tokenizer):
+class PolypeptideTruncation(PolypeptideSequenceVariantBase):
     """Class for tokenizing Protein Substitution."""
-
-    def __init__(self, amino_acid_cache: AminoAcidCache) -> None:
-        """Initialize the Protein Substitution Class.
-
-        :param AminoAcidCache amino_acid_cache: Valid amino acid codes.
-        """
-        self.amino_acid_cache = amino_acid_cache
-        self.splitter = re.compile(r'(\d+)')
-        self.psub = None
 
     def match(self, input_string) -> Optional[PolypeptideTruncationToken]:
         """Return a PolypeptideTruncationToken match if one exists.  # noqa: D202, E501
@@ -101,21 +90,3 @@ class PolypeptideTruncation(Tokenizer):
                     psub_parts = \
                         self.splitter.split(psub_parts[0].split('p.')[-1])
                     self._get_psub(psub_parts)
-
-    def _set_psub(self, amino_acid, position, new_amino_acid):
-        """Initialize protein substitution.
-
-        :param str amino_acid: Reference amino acid
-        :param str position: The position of the amino acid substituted
-        :param str new_amino_acid: The new amino_acid
-        """
-        self.psub['amino_acid'] = amino_acid
-        self.psub['position'] = int(position)
-        self.psub['new_amino_acid'] = new_amino_acid
-
-    def _is_valid_amino_acid(self, amino_acids):
-        """Return whether or not amino acids are valid."""
-        for amino_acid_code in amino_acids:
-            if not self.amino_acid_cache.__contains__(amino_acid_code):
-                return False
-        return True
